Moved constant user from oracle-e2e to e2e-common
