+++ e2e-commons/README.md
@@ -9,6 +9,11 @@ Common scripts and configuration for the end-to-end tests.
 ```
 Spins up parity networks, redis, rabbit, e2e container needed for end-to-end tests.
 
+```
+./down.sh
+```
+Shuts down and cleans up containers, networks, services, running scripts.
+
 ### Components
 
 | Component | Description |
