+++ e2e-commons/constants.json
@@ -1,0 +1,6 @@
+{
+  "user": {
+    "address": "0x7FC1442AB55Da569940Eb750AaD2BAA63DA4010E",
+    "privateKey": "0x460635eb4ac4287de2d2393985e19b4a9f948ac533453a1044ab8d50330b0df9"
+  }
+}
