+++ oracle-e2e/test/ercToNative.js
@@ -2,7 +2,8 @@ const path = require('path')
 const Web3 = require('web3')
 const assert = require('assert')
 const promiseRetry = require('promise-retry')
-const { user, contractsPath } = require('../constants.json')
+const { user } = require('../../e2e-commons/constants.json')
+const { contractsPath } = require('../constants.json')
 const { generateNewBlock } = require('../../e2e-commons/utils')
 
 const abisDir = path.join(__dirname, '..', contractsPath, 'build/contracts')
